backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
asyncpg==0.29.0
redis==5.0.1
celery==5.3.4
httpx==0.25.2
openai==1.3.7
google-generativeai==0.3.0
chromadb==0.4.18
langchain==0.0.340
PyMuPDF==1.23.8
pandas==2.1.3
numpy==1.26.2
websockets==12.0
python-socketio==5.10.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
